LayoutTests/imported/w3c/web-platform-tests/svg/text/scripted/textpath-textlength-text-anchor-001.tentative-expected.txt

XXXX
XXXX
XXXX
XXXX
XXXX
XXXX

FAIL 'textLength' on <textPath> with 'text-anchor' adjustments assert_approx_equals: width, text-anchor=start expected 25 +/- 0.5 but got 38.75
